app/requirements.txt
flask==3.0.3
azure-storage-blob==12.19.1
msal
